Load exhibit data from local disk
This feature allows the user to upload a CSV that specifies a
file instead of a url. At import time, the filename is converted
to a localhost url, and the file is delievered via a local static
file server in the rails application. From there, the normal
spotlight code takes over and the image is treated exactly like
a file imported via a url.

The directory containing the files up be uploaded must be specified
in config/secrets.yml.

A feature test for the new behavior is also included.