Add UniversalViewer via yarn
This commit changes `package.json` so that
is possible to install `universalviewer`
with yarn.

To install uv in your development
enviroment run `yarn run install-uv`. There
is a capistrano task that installs UniversalViewer
included in this commit.

On the front-end, the catalog controller is configured
to use the `universal_viewer_default` partial. This
uses the `<iframe>` version of UniversalViewer. It's
done this way because including UV directly is not
compatible with jQuery almond and other JS included
in `spotlight.js`.

Documentation on the options available for the viewer
is here:

https://github.com/UniversalViewer/universalviewer/wiki/Embedding#hash-parameters

There is a `ManifestService` and a `#iiif_manifest` helper that are
used in views. These are tested on an actual item in the existing
feature specs.

Connected to #24